Add some post-bulid messages to ./build.sh script
erljs.html and erljs_code.js may need manual updating, if some new
modules are needed. erljs.html and erljs_code.js is provided in
repository, but in future it should be removed, and compiled from
sources automatically.

In fact module will be loaded/downloaded dynamically, and dynamically
added to VM, so both of this steps will not be nacassary, but some list
of pre-loaded module can be helpful (for example, if we know we will use
lists, erlang, gen_server, gen, proplists anyway, it is good to preload
them, even in single request to server, or embeded directly somewhere).

For now, build.sh just displays small notice what needs to be manually
updated and how.