Fix 'div' operator
Apperently I not tested it well enough (and div 3 case was not
well choicen). 10 div 3 is 3, and we was calculating it using
10/3 = 3.333 and then rounding, which gives 3.

But this fails when doing div 2, for example 9/2 = 4.5, which rounded
gives 5, not 4!

One of the possibility is to use Math.floor instead of Math.round,
but then results are incorrect for negative arguments.

Just use binary operator magic, to cast result to integer. Not tested
which solution is faster ( binary or 0, right shift 0, double binary
negation), so just using what I think should be fast, but leaving
everything in comments for future investigation.

This fixes lists:seq(1,20,2), which previously returned [1,3,...,19,21],
not it correctly returns [1,3,...,19].