Move eval calls from erljs:eval/1 to separate function
eval call is evil, not only because it make more dynamic, and harder to
debug/read, but also because JavaScript engine/compiler cannot optimize
given scope (because eval can dynamically add variables to this scope,
for example). Move eval (which we probably needs, at least now) to
separate function, so bad effects of using it will be only contained in
single function.

Tested perfromance of new code, but I see no difference.